Test examples using pytest
